added user sign up and log in feature
