+++ next.config.mjs
@@ -1,4 +1,15 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    async redirects() {
+        return [
+          {
+            source: '/',
+            destination: '/login',
+            permanent: false, // Set to true if you want the redirect to be permanent
+          },
+        ];
+      },
+
+};
 
 export default nextConfig;
