+++ package.json
@@ -13,6 +13,7 @@
     "@emotion/styled": "^11.13.0",
     "@google/generative-ai": "^0.17.1",
     "@mui/material": "^5.16.7",
+    "firebase": "^10.13.0",
     "next": "^14.2.5",
     "react": "^18",
     "react-dom": "^18"
